fix typography and reorganize files
