+++ lib/phx_tools_web/live/phx_tools_live/icons.ex
@@ -10,8 +10,8 @@ defmodule PhxToolsWeb.PhxToolsLive.Icons do
   def exclamation_icon(assigns) do
     ~H"""
     <svg
-      width="20"
-      height="20"
+      width="22"
+      height="22"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 24 24"
       fill="currentColor"
